feat: logging in docker

+++ user-service/src/main/java/org/dainn/userservice/service/impl/InvitationService.java
@@ -8,10 +8,13 @@ import org.dainn.userservice.dto.user.UserDto;
 import org.dainn.userservice.event.EventProducer;
 import org.dainn.userservice.exception.AppException;
 import org.dainn.userservice.exception.ErrorCode;
+import org.dainn.userservice.jwt.JwtProvider;
 import org.dainn.userservice.mapper.IInvitationMapper;
 import org.dainn.userservice.mapper.IUserMapper;
 import org.dainn.userservice.model.Invitation;
+import org.dainn.userservice.model.User;
 import org.dainn.userservice.repository.IInvitationRepository;
+import org.dainn.userservice.repository.IUserRepository;
 import org.dainn.userservice.service.IInvitationService;
 import org.dainn.userservice.service.IUserService;
 import org.dainn.userservice.util.Paging;
@@ -21,6 +24,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -28,8 +32,10 @@ public class InvitationService implements IInvitationService {
     private final IInvitationRepository invitationRepository;
     private final IInvitationMapper invitationMapper;
     private final IUserService userService;
+    private final IUserRepository userRepository;
     private final IUserMapper userMapper;
     private final EventProducer eventProducer;
+    private final JwtProvider jwtProvider;
 
     @Transactional
     @Override
@@ -38,15 +44,21 @@ public class InvitationService implements IInvitationService {
                 .ifPresent(invitation -> {
                     throw new AppException(ErrorCode.INVITATION_EXISTED);
                 });
+        userRepository.findByEmail(dto.getEmail())
+                .ifPresent(user -> {
+                    throw new AppException(ErrorCode.EMAIL_EXISTED);
+                });
         Invitation entity = invitationMapper.toEntity(dto);
         entity = invitationRepository.save(entity);
         UserDto user = userService.findOwnerByAgency(dto.getAgencyId());
+        String token = jwtProvider.generateToken(dto.getEmail(), entity.getAgencyId());
         MailData mailData = MailData.builder()
+                .inviteId(entity.getId())
+                .from(user.getEmail())
                 .to(dto.getEmail())
                 .subject("Your Invitation to Join Web builder")
                 .inviterName(user.getName())
-                .invitationLink("http://localhost:3000")
-                .agencyId(entity.getAgencyId())
+                .invitationLink("http://localhost:3000/accept-invite?token=" + token)
                 .build();
         eventProducer.sendInviteEvent(mailData);
         return invitationMapper.toDto(invitationRepository.save(entity));
@@ -57,12 +69,20 @@ public class InvitationService implements IInvitationService {
     @Transactional
     @Override
     public String verify(String email, UserDto userInfo) {
-        Invitation invitation = invitationRepository.findByEmailAndStatus(email, InvitationStatus.PENDING)
-                .orElseThrow(() -> new AppException(ErrorCode.INVITATION_NOT_EXISTED));
-        userInfo = userMapper.toUserInfo(userInfo, invitation);
-        userService.create(userInfo);
-        invitationRepository.deleteById(invitation.getId());
-        return invitation.getAgencyId();
+        try {
+            Invitation invitation = invitationRepository.findByEmailAndStatus(email, InvitationStatus.PENDING)
+                    .orElseThrow(() -> new AppException(ErrorCode.INVITATION_NOT_EXISTED));
+            Optional<User> optional = userRepository.findByEmail(userInfo.getEmail());
+            if (optional.isPresent()) {
+                return optional.get().getAgencyId();
+            }
+            userInfo = userMapper.toUserInfo(userInfo, invitation);
+            userService.create(userInfo);
+            invitationRepository.deleteById(invitation.getId());
+            return invitation.getAgencyId();
+        } catch (Exception e) {
+            return null;
+        }
     }
 
     @Transactional
